(shmif/net) changed allocate_connpoint
This should fix the issue with creating multiple connections, and
relates to a design flaw that emerged when factoring out the shmif-
server library from the main engine code.

The culprit is in the part that allocating a connection point returns
an opaque struct from which the descriptor should be extracted before
polling, or its semantics will change from the bound socket to the
accepted one.

Ancilliary tools and tests have not been fully fixed / reworked to
account for this change yet.